Push

+++ core/src/com/kamigaku/dungeoncrawler/tile/Ground.java
@@ -8,14 +8,14 @@ import com.kamigaku.dungeoncrawler.component.GraphicsComponent;
 
 public class Ground extends Tile {
     
-    public Ground(Sprite sprite, float x, float y) {
+    public Ground(Sprite sprite, int x, int y) {
         this._graphics = new GraphicsComponent(sprite, 0f, 0f);
         this.x = x;
         this.y = y;
         this.isCrossable = true;
     }
     
-    public Ground(String sprite, float x, float y) {
+    public Ground(String sprite, int x, int y) {
         this._graphics = new GraphicsComponent(sprite, -0f, 0f);
         this.x = x;
         this.y = y;
